feat: 다시 Spring Security + Spring Session을 혼합해서 쓰는 방식으로 변경
- 튜터님께 여쭤본 끝에, 세션 리포지토리에 SecurityContext를 넣는 것이 맞는 구현이라고 판단하여 다시 롤백
- 처음엔 SecurityContext를 세션 저장소에 넣는 것이 맞는 구현인지 헷갈렸었으나, 어차피 Authentication 객체로 인증 가능, 추후에 직렬화만 잘 신경써주자.
- UserDetailsServiceImpl은 삭제
