+++ src/main/java/com/reboot_course/firstcome_system/auth/core/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package com.reboot_course.firstcome_system.auth.dto.request;
+package com.reboot_course.firstcome_system.auth.core.dto.request;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
