+++ src/main/java/com/reboot_course/firstcome_system/auth/core/service/AuthService.java
@@ -1,9 +1,6 @@
-package com.reboot_course.firstcome_system.auth.service;
+package com.reboot_course.firstcome_system.auth.core.service;
 
-import com.reboot_course.firstcome_system.auth.dto.request.LoginRequest;
-import com.reboot_course.firstcome_system.auth.dto.request.SignupRequest;
-import com.reboot_course.firstcome_system.member.entity.Member;
-import com.reboot_course.firstcome_system.member.repository.MemberRepository;
+import com.reboot_course.firstcome_system.auth.core.dto.request.LoginRequest;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpSession;
 import lombok.RequiredArgsConstructor;
@@ -13,49 +10,31 @@ import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
 import org.springframework.security.core.context.SecurityContext;
 import org.springframework.security.core.context.SecurityContextHolder;
-import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.security.web.context.HttpSessionSecurityContextRepository;
 import org.springframework.stereotype.Service;
-
-import java.time.LocalDateTime;
 
 @Service
 @RequiredArgsConstructor
 public class AuthService {
-    private final MemberRepository memberRepository;
-    private final PasswordEncoder passwordEncoder;
     private final AuthenticationManager authenticationManager;
-
-    public int signUp(SignupRequest request) {
-        if (memberRepository.findByEmail(request.getEmail()).isPresent()) {
-            throw new RuntimeException("Email is already in use");
-        }
-
-        Member member = Member.builder()
-                .name(request.getName())
-                .email(request.getEmail())
-                .password(passwordEncoder.encode(request.getPassword()))
-                .phone(request.getPhone())
-                .address(request.getAddress())
-                .lastPasswordUpdated(LocalDateTime.now())
-                .build();
-
-        Member savedMember = memberRepository.save(member);
-        return savedMember.getId();
-    }
 
     public String login(LoginRequest loginRequest, HttpServletRequest request) {
         try {
+            // 1. 인증객체 생성
             Authentication authentication = authenticationManager.authenticate(
                     new UsernamePasswordAuthenticationToken(loginRequest.getEmail(), loginRequest.getPassword())
             );
 
+            // 2. 컨텍스트 홀더에 인증객체 저장
             SecurityContext context = SecurityContextHolder.createEmptyContext();
             context.setAuthentication(authentication);
             SecurityContextHolder.setContext(context);
 
+            // 3. 위에서 만든 컨텍스트를 세션에 저장
             HttpSession session = request.getSession(true);
             session.setAttribute(HttpSessionSecurityContextRepository.SPRING_SECURITY_CONTEXT_KEY, context);
+
+            // 4. 세션 ID 반환
             return session.getId();
         } catch (AuthenticationException e) {
             throw new RuntimeException("Authentication failed", e);
@@ -63,20 +42,13 @@ public class AuthService {
     }
 
     public void logout(HttpServletRequest request) {
+        // 1. 세션에서 비활성화
         HttpSession session = request.getSession(false);
         if (session != null) {
             session.invalidate();
         }
 
+        // 2. 컨텍스트 홀더 비우기
         SecurityContextHolder.clearContext();
-    }
-
-    public String getCurrentUserEmail() {
-        Authentication authentication = SecurityContextHolder.getContext().getAuthentication();
-        if (authentication == null || !authentication.isAuthenticated()) {
-            throw new RuntimeException("No authenticated user found");
-        }
-
-        return authentication.getName();
     }
 }
