+++ src/main/java/com/reboot_course/firstcome_system/auth/security/provider/AuthenticationproviderImpl.java
@@ -1,4 +1,4 @@
-package com.reboot_course.firstcome_system.auth.security;
+package com.reboot_course.firstcome_system.auth.security.provider;
 
 import com.reboot_course.firstcome_system.member.entity.Member;
 import com.reboot_course.firstcome_system.member.repository.MemberRepository;
@@ -8,16 +8,20 @@ import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.AuthenticationException;
+import org.springframework.security.core.GrantedAuthority;
+import org.springframework.security.core.authority.SimpleGrantedAuthority;
+import org.springframework.security.core.userdetails.User;
+import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.security.core.userdetails.UsernameNotFoundException;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Component;
 
 import java.util.Collections;
+import java.util.List;
 
 @Component
 @RequiredArgsConstructor
-public class CustomAuthenticationProvider implements AuthenticationProvider {
-
+public class AuthenticationproviderImpl implements AuthenticationProvider {
     private final MemberRepository memberRepository;
     private final PasswordEncoder passwordEncoder;
 
@@ -33,7 +37,11 @@ public class CustomAuthenticationProvider implements AuthenticationProvider {
             throw new BadCredentialsException("Invalid password");
         }
 
-        return new UsernamePasswordAuthenticationToken(email, password, Collections.emptyList());
+        List<GrantedAuthority> authorities = Collections.singletonList(new SimpleGrantedAuthority("ROLE_USER"));
+
+        UserDetails userDetails = new User(member.getEmail(), member.getPassword(), authorities);
+
+        return new UsernamePasswordAuthenticationToken(userDetails, null, authorities);
     }
 
     @Override
