+++ src/main/java/com/reboot_course/firstcome_system/auth/session/config/SessionConfig.java
@@ -1,4 +1,4 @@
-package com.reboot_course.firstcome_system.auth.config.session;
+package com.reboot_course.firstcome_system.auth.session.config;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -6,6 +6,8 @@ import org.springframework.session.MapSessionRepository;
 import org.springframework.session.config.annotation.web.http.EnableSpringHttpSession;
 import org.springframework.session.web.http.CookieSerializer;
 import org.springframework.session.web.http.DefaultCookieSerializer;
+import org.springframework.session.web.http.HeaderHttpSessionIdResolver;
+import org.springframework.session.web.http.HttpSessionIdResolver;
 
 import java.util.concurrent.ConcurrentHashMap;
 
@@ -16,6 +18,11 @@ public class SessionConfig {
     @Bean
     public MapSessionRepository sessionRepository() {
         return new MapSessionRepository(new ConcurrentHashMap<>());
+    }
+
+    @Bean
+    public HttpSessionIdResolver httpSessionIdResolver() {
+        return new HeaderHttpSessionIdResolver("X-Auth-Token");
     }
 
     @Bean
